refactor: mover implementacion de factory a infrastructura port/in
