+++ src/main/java/com/teamcubation/api/pacientes/infrastructure/adapter/in/exporter/factory/ExporterFactory.java
@@ -1,0 +1,7 @@
+package com.teamcubation.api.pacientes.infrastructure.adapter.in.exporter.factory;
+
+import com.teamcubation.api.pacientes.infrastructure.adapter.in.exporter.PatientExporter;
+
+public interface ExporterFactory {
+    PatientExporter createExporter();
+}
